third_party/blink/web_tests/clipboard/async-clipboard/async-write-text-read-blobtext.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard writeText -> read ([text/plain Blob]) tests</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../http/tests/resources/permissions-helper.js"></script>
<script>
async function readWriteTest(textInput) {
  promise_test(async t => {
    await PermissionsHelper.setPermission('clipboard-read', 'granted');
    await PermissionsHelper.setPermission('clipboard-write', 'granted');

    await navigator.clipboard.writeText(textInput);
    const blobsOutput = await navigator.clipboard.read();
    assert_equals(Object.keys(blobsOutput).length, 1);
    const blobOutput = blobsOutput['text/plain'];
    assert_equals(blobOutput.type, 'text/plain');

    const textOutput = await (new Response(blobOutput)).text();
    assert_equals(textOutput, textInput);
  }, 'Verify write and read clipboard given text: ' + textInput);
}

readWriteTest('Clipboard write text -> read ([text/plain Blob]) test');
readWriteTest('non-Latin1 text encoding test データ');
</script>
